rename sections to layout-sections
https://coveord.atlassian.net/browse/KIT-5096
